\name{finebalance-package}
\alias{finebalance-package}
\docType{package}
\title{
Approximate fine balance when exact fine balance is not achievable
}
\description{
Uses the algorithm of minimum cost network flow.
}
\details{
\tabular{ll}{
Package: \tab finebalance\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-06-29\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
}
\author{
Dan Yang <danyang@wharton.upenn.edu>
}
\references{
Optimal matching with minimal deviation from fine balance in a study of obesity and surgical outcomes
}
\keyword{nonparametric}
\keyword{network optimization}
\keyword{multivariate matching}
